pyproject.toml
[project]
name = "calli-cli"
version = "0.1.0"
description = ""
authors = [
    {name = "Arlo Purcell",email = "arlo@purcell.live"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial (>=3.5,<4.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "hershey-fonts (>=2.1.0,<3.0.0)"
]

[project.scripts]
calli = 'calli_cli.main:main'

[tool.poetry]
packages = [{include = "calli_cli", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
